fix: return 404 for missing import jobs (#2156)

+++ webapp/src/main/java/io/github/microcks/web/JobController.java
@@ -155,8 +155,13 @@ public class JobController {
    public ResponseEntity<ImportJob> activateJob(@PathVariable("id") String jobId, UserInfo userInfo) {
       log.debug("Activating job with id {}", jobId);
       ImportJob job = jobRepository.findById(jobId).orElse(null);
-      if (job != null && (authorizationChecker.hasRole(userInfo, AuthorizationChecker.ROLE_ADMIN)
-            || authorizationChecker.hasRoleForImportJob(userInfo, AuthorizationChecker.ROLE_MANAGER, job))) {
+
+      if (job == null) {
+         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+      }
+
+      if (authorizationChecker.hasRole(userInfo, AuthorizationChecker.ROLE_ADMIN)
+            || authorizationChecker.hasRoleForImportJob(userInfo, AuthorizationChecker.ROLE_MANAGER, job)) {
          job.setActive(true);
          initMetadataIfMissing(job);
          job.getMetadata().objectUpdated();
@@ -208,6 +213,11 @@ public class JobController {
    public ResponseEntity<String> deleteJob(@PathVariable("id") String jobId, UserInfo userInfo) {
       log.debug("Removing job with id {}", jobId);
       ImportJob job = jobRepository.findById(jobId).orElse(null);
+
+      if (job == null) {
+         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+      }
+
       if (authorizationChecker.hasRole(userInfo, AuthorizationChecker.ROLE_ADMIN)
             || authorizationChecker.hasRoleForImportJob(userInfo, AuthorizationChecker.ROLE_MANAGER, job)) {
          jobRepository.deleteById(jobId);
